+++ src/services/matrix.js
@@ -12,10 +12,9 @@ import request from 'xmlhttp-request';
 import AsyncStorage from '@react-native-community/async-storage';
 import AsyncCryptoStore from '../storage/AsyncCryptoStore';
 import Olm from 'olm/olm_legacy';
-//import { generateSecureRandom } from 'react-native-securerandom';
+import { toImageBuffer } from '../utilities/misc';
 
 // import i18n from '../../i18n';
-// import { toImageBuffer } from '../../utilities';
 // import SqlStore from './SqlStore';
 
 const debug = require('debug')('rnm:matrix.js');
@@ -159,6 +158,7 @@ class MatrixService {
         break;
       case 'ERROR':
         debug('A syncing error ocurred:', { state, prevState, data });
+        console.warn('A syncing error ocurred:', { state, prevState, data });
         this._isSynced$.next(false);
         this._error$.next(data);
         break;
@@ -223,17 +223,17 @@ class MatrixService {
   }
 
   async uploadImage(image) {
-    // try {
-    //   const url = await this._client.uploadContent(toImageBuffer(image.data), {
-    //     onlyContentUri: true,
-    //     name: image.fileName,
-    //     type: image.type,
-    //   });
-    //   return url;
-    // } catch (e) {
-    //   debug('Error uploading image:', e);
-    //   return null;
-    // }
+    try {
+      const url = await this._client.uploadContent(toImageBuffer(image.data), {
+        onlyContentUri: true,
+        name: image.fileName,
+        type: image.type,
+      });
+      return url;
+    } catch (e) {
+      debug('Error uploading image:', e);
+      return null;
+    }
   }
 }